+++ packages/core/minos-microservice-aggregate/minos/aggregate/__init__.py
@@ -1,6 +1,6 @@
 __author__ = "Minos Framework Devs"
 __email__ = "hey@minos.run"
-__version__ = "0.5.2"
+__version__ = "0.5.3"
 
 from .actions import (
     Action,
@@ -42,6 +42,7 @@ from .exceptions import (
     EventRepositoryConflictException,
     EventRepositoryException,
     NotFoundException,
+    RefException,
     SnapshotRepositoryConflictException,
     SnapshotRepositoryException,
     TransactionNotFoundException,