+++ packages/core/minos-microservice-aggregate/minos/aggregate/events/models.py
@@ -3,6 +3,7 @@ from __future__ import (
 )
 
 import logging
+import warnings
 from datetime import (
     datetime,
 )
@@ -64,18 +65,38 @@ class Event(DeclarativeModel):
     def __lt__(self, other: Any) -> bool:
         return isinstance(other, type(self)) and self.version < other.version
 
-    def __getattr__(self, item: str) -> Any:
+    def __getitem__(self, item: str) -> Any:
         try:
-            return super().__getattr__(item)
-        except AttributeError as exc:
+            return super().__getitem__(item)
+        except KeyError as exc:
             if item != "fields_diff":
                 try:
-                    return self.get_one(item)
+                    return self.get_field(item)
                 except Exception:
                     raise exc
             raise exc
 
+    def __getattr__(self, item: str) -> Any:
+        try:
+            return super().__getattr__(item)
+        except AttributeError as exc:
+            try:
+                return self[item]
+            except Exception:
+                raise exc
+
     def get_one(self, name: str, return_diff: bool = False) -> Union[FieldDiff, Any, list[FieldDiff], list[Any]]:
+        """Get first field diff with given name.
+
+        :param name: The name of the field diff.
+        :param return_diff: If ``True`` the result is returned as field diff instances, otherwise the result is
+            returned as value instances.
+        :return: A ``FieldDiff`` instance.
+        """
+        warnings.warn("get_one() method is deprecated by get_field() and will be removed soon.", DeprecationWarning)
+        return self.get_field(name, return_diff)
+
+    def get_field(self, name: str, return_diff: bool = False) -> Union[FieldDiff, Any, list[FieldDiff], list[Any]]:
         """Get first field diff with given name.
 
         :param name: The name of the field diff.
@@ -86,6 +107,16 @@ class Event(DeclarativeModel):
         return self.fields_diff.get_one(name, return_diff)
 
     def get_all(self, return_diff: bool = False) -> dict[str, Union[FieldDiff, Any, list[FieldDiff], list[Any]]]:
+        """Get all field diffs with given name.
+
+        :param return_diff: If ``True`` the result is returned as field diff instances, otherwise the result is
+            returned as value instances.
+        :return: A list of ``FieldDiff`` instances.
+        """
+        warnings.warn("get_all() method is deprecated by get_fields() and will be removed soon.", DeprecationWarning)
+        return self.get_fields(return_diff)
+
+    def get_fields(self, return_diff: bool = False) -> dict[str, Union[FieldDiff, Any, list[FieldDiff], list[Any]]]:
         """Get all field diffs with given name.
 
         :param return_diff: If ``True`` the result is returned as field diff instances, otherwise the result is